+++ OptiScaler/shaders/resource_flip/RF_Dx12.cpp
@@ -62,7 +62,7 @@ static bool CreateComputeShader(ID3D12Device* device, ID3D12RootSignature* rootS
 }
 
 bool RF_Dx12::Dispatch(ID3D12Device* InDevice, ID3D12GraphicsCommandList* InCmdList, ID3D12Resource* InResource,
-                       ID3D12Resource* OutResource, UINT width, UINT height, bool velocity)
+                       ID3D12Resource* OutResource, UINT64 width, UINT height, bool velocity)
 {
     if (!_init || InDevice == nullptr || InCmdList == nullptr || InResource == nullptr || OutResource == nullptr)
         return false;
@@ -127,7 +127,7 @@ bool RF_Dx12::Dispatch(ID3D12Device* InDevice, ID3D12GraphicsCommandList* InCmdL
     RFConstants constants {};
 
     constants.height = height - 1;
-    constants.width = width - 1;
+    constants.width = static_cast<UINT>(width - 1);
     constants.offset = Config::Instance()->FGResourceFlipOffset.value_or_default() ? inDesc.Height - height : 0;
     constants.velocity = velocity ? 1 : 0;
